@@ -2,7 +2,7 @@
   "name": "the-red-planet",
   "version": "1.0.0",
   "description": "The Red Planet - Survival Game",
-  "main": "main.js",
+  "main": "main.mjs",
   "scripts": {
     "start": "electron ."
   },
@@ -14,6 +14,6 @@
   "author": "Francis Stuidos",
   "license": "",
   "devDependencies": {
-    "electron": "^5.0.0"
+    "electron": "^30.0.0"
   }
 }
